move rfuse parsing into file parsing
